requirements.txt
python-dotenv
requests
python-telegram-bot
pyyaml
schedule